requirements.txt
langchain
langgraph
langchain-community
langchain-google-genai
langchain-ollama
chromadb
pypdf
pdf2image
pillow
streamlit
python-dotenv
requests
sentence-transformers
